+++ src/lib/config.js
@@ -10,7 +10,7 @@ const defaultConfig = {
 	cookieDomain: null,
 	customPurposeListLocation: null, // './purposes.json',
 	// The location of the latest vendorlist to use.
-	globalVendorListLocation: 'https://vendorlist.consensu.org/vendorlist.json',
+	globalVendorListLocation: `https://s.flocdn.com/cmp/${VERSION}/config/vendorlist.json`,
 	globalConsentLocation: './portal.html',
 	pubVendorListLocation: null,
 	storeConsentGlobally: false,
@@ -31,6 +31,10 @@ class Config {
 	}
 
 	update = (updates) => {
+		if (updates && updates.globalVendorListLocation === 'https://vendorlist.consensu.org/vendorlist.json') {
+			delete updates.globalVendorListLocation;
+		}
+
 		if (updates && typeof updates === 'object') {
 			const validKeys = Object.keys(defaultConfig);
 			const { validUpdates, invalidKeys } = Object.keys(updates).reduce(