+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## [1.5.7](https://github.com/openmail/system1-cmp/compare/v1.5.6...v1.5.7) (2020-08-17)
+
+### Fix
+
+- [x] Default to hosted vendorlist since IAB removed their global v1 vendorlist
+- [x] Check for IAB global vendor list location and remove if manually set
+- [x] Add SameSite;secure cookie props for http and https
+
 ## [1.5.6](https://github.com/openmail/system1-cmp/compare/v1.5.5...v1.5.6) (2020-06-23)
 
 ### Translations